contact list,details,delete completed

+++ EmailMarketing.Framework/Services/Contacts/ContactService.cs
@@ -1,4 +1,5 @@
-﻿using EmailMarketing.Common.Extensions;
+﻿using EmailMarketing.Common.Exceptions;
+using EmailMarketing.Common.Extensions;
 using EmailMarketing.Framework.Entities.Contacts;
 using EmailMarketing.Framework.UnitOfWork.Contacts;
 using Microsoft.EntityFrameworkCore;
@@ -22,30 +23,31 @@ namespace EmailMarketing.Framework.Services.Contacts
         public async Task<(IList<Contact> Items, int Total, int TotalFilter)> GetAllContactAsync(
             Guid? userId,string searchText, string orderBy, int pageIndex, int pageSize)
         {
-
-            //var result =  await _contactUnitOfWork.ContactRepository.GetAsync(x => x,
-            //     x => !x.IsDeleted && x.IsActive && (!userId.HasValue || x.Group.UserId == userId.Value),
-            //     x => x.OrderByDescending(o => o.Created),
-            //     x => x.Include(o => o.Group).Include(o => o.ContactValueMaps).ThenInclude(o => o.FieldMap),
-            //     true);
-
             var columnsMap = new Dictionary<string, Expression<Func<Entities.Contacts.Contact, object>>>()
             {
                 ["Email"] = v => v.Email
             };
             var result = await _contactUnitOfWork.ContactRepository.GetAsync<Entities.Contacts.Contact>(
-                x => x, x => x.UserId == userId,
-                x => x.ApplyOrdering(columnsMap, orderBy), null,
+                x => x, x => (!userId.HasValue || x.UserId == userId.Value)&& (x.Email.Contains(searchText)),
+                x => x.ApplyOrdering(columnsMap, orderBy), x=> x.Include(i=> i.ContactGroups).ThenInclude(i=>i.Group),
                 pageIndex, pageSize, true);
             return (result.Items, result.Total, result.TotalFilter);
         }
-        public async Task<Entities.Contacts.Contact> GetByIdAsync(int id)
+        public async Task<Contact> GetByIdAsync(int id)
         {
-            return await _contactUnitOfWork.ContactRepository.GetByIdAsync(id);
+            var result = await _contactUnitOfWork.ContactRepository.GetFirstOrDefaultAsync(
+                x => x, x => x.Id == id,
+                x => x.Include(i => i.ContactGroups).ThenInclude(i => i.Group)
+                        .Include(i => i.ContactValueMaps).ThenInclude(i => i.FieldMap));
+
+            if (result == null) throw new NotFoundException(nameof(Contact), id);
+
+            return result;
         }
-        public async Task<Entities.Contacts.Contact> DeleteAsync(int id)
+        public async Task<Contact> DeleteAsync(int id)
         {
             var contact = await GetByIdAsync(id);
+            if (contact == null) throw new NotFoundException(nameof(Contact), id);
             await _contactUnitOfWork.ContactRepository.DeleteAsync(id);
             await _contactUnitOfWork.SaveChangesAsync();
             return contact;
